Add files via upload
Theory:  
Simple Columnar Technique:
The message is written out in rows of a fixed length, and then read out again column by column, and the columns are chosen in some scrambled order.
Width of the rows and the permutation of the columns are usually defined by a keyword. For example, the word HACK is of length 4 (so the rows are of length 4), and the permutation is defined by the alphabetical order of the letters in the keyword. In this case, the order would be “3 1 2 4”. Any spare spaces are filled with nulls or left blank or placed by Filler alphabets Finally, the message is read off in columns, in the order specified by the numbers. 1,2,3,4.