Add files via upload
Theory:  
RSA:
RSA algorithm is a public key encryption technique and is considered as the most secure way of encryption. It was invented by Rivest, Shamir and Adleman in year 1978 and hence name RSA algorithm. The entire RSA is all about creating public key, private key for encryption and decryption