Add files via upload
Theory:  
Message Digest is used to ensure the integrity of a message transmitted over an insecure channel (where the content of the message can be changed). The message is passed through a Cryptographic hash function. This function creates a compressed image of the message called Digest.
Produces 128-bit message digest.
Let’s assume, Alice sent a message and digest pair to Bob. To check the integrity of the message Bob runs the cryptographic hash function on the received message and gets a new digest. Now, Bob will compare the new digest and the digest sent by Alice. If, both are same then Bob is sure that the original message is not changed.
 
Steps for MD5:
1. Append Padding bits - convert data bits into multiples of 512.
2. Append length bits - dividing padded bits into 512 blocks of each
3. Initialize MD buffer - create a space ready for processing of the data blocks
4. Process each 512-bit block - 16 rounds with 4 operations in each round. (16*4 = 64 operations)
5. Output message digest - the final digest created.


MD5 Operations:
The operations are:
1. Rotation - moving data from its original location to other.
2. ANDing - logical AND operation.
3. ORing - logical OR operation.
4. Substitution - substituting original values with nuetral values.